RavenDB-24723 add better validation
