+++ src/Raven.Server/Documents/Handlers/AI/Agents/AiAgentProcessorForAddOrUpdateAiAgent.cs
@@ -1,19 +1,14 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using JetBrains.Annotations;
 using Raven.Client.Documents.Operations.AI;
 using Raven.Client.Documents.Operations.AI.Agents;
 using Raven.Client.Json.Serialization;
-using Raven.Server.Documents.AI;
+using Raven.Server.Config.Categories;
 using Raven.Server.Documents.Handlers.Processors;
-using Raven.Server.Documents.Queries;
-using Raven.Server.Documents.Queries.AST;
 using Raven.Server.ServerWide.Commands.AI;
 using Sparrow.Json;
 using Sparrow.Json.Parsing;
-using Sparrow.Server.Json.Sync;
 
 namespace Raven.Server.Documents.Handlers.AI.Agents;
 internal class AiAgentProcessorForAddOrUpdateAiAgent<TRequestHandler, TOperationContext> : AbstractDatabaseHandlerProcessor<TRequestHandler, TOperationContext>
@@ -39,7 +34,8 @@ internal class AiAgentProcessorForAddOrUpdateAiAgent<TRequestHandler, TOperation
         if (string.IsNullOrEmpty(cfg.Identifier))
             cfg.Identifier = EmbeddingsGenerationConfiguration.GenerateIdentifier(cfg.Name);
         
-        ValidateConfiguration(context, cfg);
+        AiAgentHelpers.AddDefaultValues(cfg, RequestHandler.Configuration.Ai);
+        AddOrUpdateAiAgentCommand.ValidateConfiguration(context, cfg);
 
         var r = await ServerStore.SendToLeaderAsync(new AddOrUpdateAiAgentCommand(RequestHandler.DatabaseName, cfg, RequestHandler.GetRaftRequestIdFromQuery()),
             token.Token);
@@ -59,75 +55,23 @@ internal class AiAgentProcessorForAddOrUpdateAiAgent<TRequestHandler, TOperation
             context.Write(writer, json);
         }
     }
+}
 
-    private void ValidateConfiguration(JsonOperationContext context, AiAgentConfiguration configuration)
+public static class AiAgentHelpers
+{
+    public static void AddDefaultValues(AiAgentConfiguration configuration, AiConfiguration aiConfig)
     {
         var reduction = configuration.ChatTrimming;
-        if (reduction != null)
+        if (reduction?.Tokens != null)
         {
-            if ((reduction.Tokens != null) == (reduction.Truncate != null))
-            {
-                throw new InvalidOperationException($"{nameof(configuration.ChatTrimming)} requires exactly one strategy: " +
-                                                    $"either {nameof(reduction.Tokens)} or {nameof(reduction.Truncate)}, " +
-                                                    "but not both or neither.");
-            }
+            if (string.IsNullOrEmpty(reduction.Tokens.SummarizationTaskBeginningPrompt))
+                reduction.Tokens.SummarizationTaskBeginningPrompt = aiConfig.SummarizationTaskBeginningPrompt;
 
-            if (reduction.Tokens != null)
-            {
-                var aiConfig = RequestHandler.Configuration.Ai;
+            if (string.IsNullOrEmpty(reduction.Tokens.SummarizationTaskEndPrompt))
+                reduction.Tokens.SummarizationTaskEndPrompt = aiConfig.SummarizationTaskEndPrompt;
 
-                if (string.IsNullOrEmpty(reduction.Tokens.SummarizationTaskBeginningPrompt))
-                    reduction.Tokens.SummarizationTaskBeginningPrompt = aiConfig.SummarizationTaskBeginningPrompt;
-
-                if (string.IsNullOrEmpty(reduction.Tokens.SummarizationTaskEndPrompt))
-                    reduction.Tokens.SummarizationTaskEndPrompt = aiConfig.SummarizationTaskEndPrompt;
-
-                if (string.IsNullOrEmpty(reduction.Tokens.ResultPrefix))
-                    reduction.Tokens.ResultPrefix = aiConfig.SummarizationResultPrefix;
-            }
-
-            if (reduction.Truncate != null)
-            {
-                var after = reduction.Truncate.MessagesLengthAfterTruncate;
-                var before = reduction.Truncate.MessagesLengthBeforeTruncate;
-                if (after > before)
-                    throw new InvalidOperationException(
-                        $"{nameof(reduction.Truncate.MessagesLengthAfterTruncate)} ({after}) must be less of equal then {nameof(reduction.Truncate.MessagesLengthBeforeTruncate)} ({before})");
-
-                if(after <= 0)
-                    throw new InvalidOperationException(
-                        $"{nameof(reduction.Truncate.MessagesLengthAfterTruncate)} ({after}) must be greater then 0");
-
-                if (before <= 0)
-                    throw new InvalidOperationException(
-                        $"{nameof(reduction.Truncate.MessagesLengthBeforeTruncate)} ({before}) must be greater then 0");
-            }
+            if (string.IsNullOrEmpty(reduction.Tokens.ResultPrefix))
+                reduction.Tokens.ResultPrefix = aiConfig.SummarizationResultPrefix;
         }
-
-        var scopeParams = configuration.Parameters;
-        foreach (var tool in configuration.Queries)
-        {
-            var q = QueryMetadata.ParseQuery(tool.Query, QueryType.Select);
-            var queryParams = new HashSet<string>(q.Parameters.Select(x => x.Value));
-            queryParams.ExceptWith(scopeParams);
-
-            string paramsSchema = ChatCompletionClient.GetSchemaForTool(tool.ParametersSchema, tool.ParametersSampleObject);
-            var schema = context.Sync.ReadForMemory(paramsSchema, "tool-schema");
-            if (schema.TryGet(ChatCompletionClient.Constants.JsonSchemaFields.Required, out BlittableJsonReaderArray required))
-            {
-                foreach (var arg in required)
-                {
-                    string queryArg = arg.ToString();
-                    if (scopeParams.Contains(queryArg))
-                        throw new InvalidOperationException($"Parameter {queryArg} is defined on both the agent level and the query level for {tool.Name}");
-
-                    queryParams.Remove(queryArg);
-                }
-            }
-
-            if (queryParams.Count > 0)
-                throw new InvalidOperationException(
-                    $"Tool query '{tool.Name}' contains parameters that are not defined in the agent configuration: '{string.Join(", ", queryParams)}'");
-        }
-    }
+    }   
 }
